requirements-gpu.txt
tqdm==4.50.2
numpy==1.19.4
tensorflow-gpu==2.1.0
tensorflow-hub
torch==1.5.0
bert-for-tf2==0.14.6
transformers==2.10.0
gensim
sklearn
seaborn
animatplot
zarr
pathos
parse
textdistance
spacy
